🎉 version(v1.2.0): 更新版本
1. 重构增益模块, 并更新大量增益效果.
2. 更新大量焚影内置战斗逻辑.
3. 更新若干lua交互, 修复若干bug. (详情查看dev分支commit记录)

+++ source/src/frame/global/skillevent.cpp
@@ -40,8 +40,8 @@ void SkillEventManager::add(int ID) {
     skillevent.SkillLevel   = std::stoi(skillevent.tab["SkillLevel"]);
     skillevent.SkillCaster  = m_ct.at(skillevent.tab["SkillCaster"]);
     skillevent.SkillTarget  = m_ct.at(skillevent.tab["SkillTarget"]);
-    skillevent.EventMask1   = std::stoi(skillevent.tab["EventMask1"]);
-    skillevent.EventMask2   = std::stoi(skillevent.tab["EventMask2"]);
+    skillevent.EventMask1   = static_cast<uint32_t>(std::stoull(skillevent.tab["EventMask1"]));
+    skillevent.EventMask2   = static_cast<uint32_t>(std::stoull(skillevent.tab["EventMask2"]));
     skillevent.EventSkillID = std::stoi(skillevent.tab["EventSkillID"]);
     // 将 SkillEvent 存入缓存
     data[ID]                = std::move(skillevent);
